Reimplement logback conf file generation, close gatling/gatling#3946
Motivation:

It was reported that the automatic logback conf file generation wasn't working as expected.
Current implementation has several flaws.

Modifications:

* Drop LogbackConfigTaskAction task as we want to get the generated file location when computing the System props for the fork process.
* Do nothing if we find one of the standard config file logback expects in the build dir.
* Otherwise, generate a file with a plugin specific name and pass it to the forked process thanks to the standard logback System property.

Result:

Use the user-defined logback.conf file, otherwise generate one in a way that wouldn't conflict with the resources.
