+++ src/main/groovy/io/gatling/gradle/GatlingRunTask.groovy
@@ -26,6 +26,38 @@ class GatlingRunTask extends DefaultTask implements JvmConfigurable {
         return project.sourceSets.gatling.scala.matching(simulationFilter)
     }
 
+    static def template(String logLevel) {
+        """<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{HH:mm:ss.SSS} [%-5level] %logger{15} - %msg%n%rEx</pattern>
+        </encoder>
+        <immediateFlush>false</immediateFlush>
+    </appender>
+    <root level="${logLevel}">
+       <appender-ref ref="CONSOLE" />
+    </root>
+</configuration>"""
+    }
+
+    private File createMissingLogbackFile() {
+        def gatlingExt = project.extensions.getByType(GatlingPluginExtension)
+        def buildDir = project.buildDir
+        def logbackXml = new File(buildDir, "resources/gatling/logback.xml")
+        def logbackTestXml = new File(buildDir, "resources/gatling/logback-test.xml")
+        def logbackGroovyXml = new File(buildDir, "resources/gatling/logback.groovy")
+
+        if (logbackXml.exists() || logbackTestXml.exists() || logbackGroovyXml.exists()) {
+            return null
+        } else {
+            def file = new File(project.buildDir, "resources/gatling/logback-gatling-gradle-plugin.xml")
+            file.parentFile.mkdirs()
+            file.text = template(gatlingExt.logLevel)
+            return file
+        }
+    }
+
     List<String> createGatlingArgs() {
 
         File scalaClasses = project.sourceSets.gatling.output.classesDirs.filter {
@@ -51,6 +83,8 @@ class GatlingRunTask extends DefaultTask implements JvmConfigurable {
     void gatlingRun() {
         def gatlingExt = project.extensions.getByType(GatlingPluginExtension)
 
+        def missingLogbackFile = createMissingLogbackFile()
+
         Map<String, ExecResult> results = simulationFilesToFQN().collectEntries { String simulationClzName ->
             [(simulationClzName): project.javaexec({ JavaExecSpec exec ->
                 exec.main = GatlingPluginExtension.GATLING_MAIN_CLASS
@@ -59,6 +93,9 @@ class GatlingRunTask extends DefaultTask implements JvmConfigurable {
                 exec.jvmArgs this.jvmArgs ?: gatlingExt.jvmArgs
                 exec.systemProperties System.properties
                 exec.systemProperties this.systemProperties ?: gatlingExt.systemProperties
+                if (missingLogbackFile != null) {
+                    exec.systemProperties(["logback.configurationFile": missingLogbackFile.absolutePath])
+                }
 
                 exec.args this.createGatlingArgs()
                 exec.args "-s", simulationClzName
@@ -74,6 +111,5 @@ class GatlingRunTask extends DefaultTask implements JvmConfigurable {
         if (!failed.isEmpty()) {
             throw new TaskExecutionException(this, new RuntimeException("There're failed simulations: ${failed.keySet().sort().join(", ")}"))
         }
-
     }
 }
